+++ packages/@cldcvr/flow-form-builder/src/components/f-form-builder/fields/emoji.ts
@@ -2,6 +2,7 @@ import { html } from "lit";
 import { FFormInputElements, FormBuilderField, FormBuilderEmojiField } from "../../../types";
 import { Ref, ref } from "lit/directives/ref.js";
 import { ifDefined } from "lit-html/directives/if-defined.js";
+import { unsafeHTML } from "lit/directives/unsafe-html.js";
 export default function (
 	name: string,
 	_field: FormBuilderField,
@@ -65,7 +66,7 @@ export default function (
 				: ""}
 			${field.helperText
 				? html`<f-div slot="help" data-qa-help-for=${field.qaId || field.id}
-						>${field.helperText}</f-div
+						>${unsafeHTML(field.helperText)}</f-div
 				  >`
 				: ``}
 			${field.label?.iconTooltip